# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Mini Banking Platform API
  description: |
    A simplified banking platform API that handles financial transactions with
    double-entry bookkeeping. The system maintains data integrity through ledger
    validation and provides user-friendly error responses following RFC 7807.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: |
        Creates a new user with initial accounts:
        - 1 USD account with $1000.00 initial balance
        - 1 EUR account with €500.00 initial balance
        
        Initial balances are funded from the system cashbook to maintain
        double-entry bookkeeping integrity.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/validation-error"
                title: "Validation Error"
                status: 400
                detail: "Email field is required"
                instance: "/auth/register"
        '409':
          description: User already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/user-already-exists"
                title: "User Already Exists"
                status: 409
                detail: "User with email user@example.com already exists"
                instance: "/auth/register"
                email: "user@example.com"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate user
      description: Authenticates a user and returns a JWT token.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://minibankingplatform.com/problems/invalid-credentials"
                title: "Invalid Credentials"
                status: 401
                detail: "The provided email or password is incorrect"
                instance: "/auth/login"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoints
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          x-oapi-codegen-extra-tags:
            validate: "required,email"
        password:
          type: string
          minLength: 8
          example: "securePassword123"
          x-oapi-codegen-extra-tags:
            validate: "required,min=8"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
          x-oapi-codegen-extra-tags:
            validate: "required,email"
        password:
          type: string
          example: "securePassword123"
          x-oapi-codegen-extra-tags:
            validate: "required"
...
    AuthResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        token:
          type: string
          description: JWT token for authentication
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: |
            A URI reference that identifies the problem type.
            This specification encourages that, when dereferenced,
            it provide human-readable documentation for the problem type.
          example: "https://minibankingplatform.com/problems/insufficient-funds"
        title:
          type: string
          description: |
            A short, human-readable summary of the problem type.
            It SHOULD NOT change from occurrence to occurrence of the problem.
          example: "Insufficient Funds"
        status:
          type: integer
          description: |
            The HTTP status code generated by the origin server for this
            occurrence of the problem.
          example: 400
        detail:
          type: string
          description: |
            A human-readable explanation specific to this occurrence
            of the problem.
          example: "Account has insufficient funds for this transfer"
        instance:
          type: string
          format: uri-reference
          description: |
            A URI reference that identifies the specific occurrence of
            the problem. It may or may not yield further information
            if dereferenced.
          example: "/transactions/transfer"
      additionalProperties: true
      example:
        type: "https://minibankingplatform.com/problems/insufficient-funds"
        title: "Insufficient Funds"
        status: 400
        detail: "Account has insufficient funds for this transfer"
        instance: "/transactions/transfer"
        available: "500.00"
        required: "1000.00"
        currency: "USD"